split testhelpers/steps.go into parts
